Finished Hunter Heidenreich Tutorial
I skipped some of the final tutorials that change the background and tiling, because that wasn't relevant for my needs. There was a strange error that occurred at some point where hunger and happiness stop updating properly. I'm not sure when that occurred, but maybe digging through older commits might illuminate the issue. I'm moving on to a different tutorial and will update this repo accordingly.